requirements.txt
pyautogui
tkinter 
pillow
pyinstaller